donnywilliams.me
These are the files for the website found at https://donnywilliams.me.

I started with a bootstrap template, then changed a lot of it to personalize the site. I took an HTML class in 1999, and I've never learned CSS or JavaScript, so I definitely had to figure some things out as I went along. But that's the point of all of this, right? We start with a certain amount of knowledge (that we're always building on), then we learn how to look up what we don't know to complete the tasks in front of us. Since this is the first real website I've ever built, there was definitely a lot of that with this. So far, nearly everything I've learned is back-end code through Python. Since I aspire to also learn how to do a lot of front-end work, it was fun to at least try my hand on some basic things with this website.

I also used a fair amount of Photoshop to get the images exactly how I wanted.